+++ terra/src/main/scala/com/dafttech/terra/engine/renderer/SubtileRendererMask.scala
@@ -18,10 +18,10 @@ class SubtileRendererMask extends SubtileRenderer {
     if (!render.isTileIndependent && render.getTile != null) {
       val offset: Vector2 = render.getTile.getRenderer.getOffset
       if (offset != null) {
-        offX = offset.x * BLOCK_SIZE
-        offY = offset.y * BLOCK_SIZE
+        offX = offset.x.toFloat * BLOCK_SIZE
+        offY = offset.y.toFloat * BLOCK_SIZE
       }
     }
-    screen.batch.draw(render.getImage, screenVec.x + offX, screenVec.y + offY, 1, 1, BLOCK_SIZE, BLOCK_SIZE, 1, 1, rotation)
+    screen.batch.draw(render.getImage, screenVec.x.toFloat + offX, screenVec.y.toFloat + offY, 1, 1, BLOCK_SIZE, BLOCK_SIZE, 1, 1, rotation)
   }
 }
