+++ terra/src/main/scala/com/dafttech/terra/game/world/items/Item.java
@@ -37,7 +37,7 @@ public abstract class Item extends GameObject implements IDrawableInventory {
 
     @Override
     public void drawInventory(Vector2 pos, AbstractScreen screen) {
-        screen.batch().draw(getImage(), pos.x() + 4, pos.y() + 4 + 12 * (1 - (getImage().getRegionHeight() / (float) getImage().getRegionWidth())), 24,
+        screen.batch().draw(getImage(), (float) pos.x() + 4, (float) pos.y() + 4 + 12 * (1 - (getImage().getRegionHeight() / (float) getImage().getRegionWidth())), 24,
                 24 * (getImage().getRegionHeight() / (float) getImage().getRegionWidth()));
     }
